project/predictive_backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
xgboost==2.1.4
pandas==2.2.3
scikit-learn==1.4.2
joblib==1.4.2
numpy==1.26.3
python-dotenv==1.0.1
